+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 
 ### IMPROVEMENTS
 
-- `[blocksync]` Make minimum receive rate configurable via `recv_rate` in `[blocksync]` config section. Operators can tune this value based on their network's block size and block time characteristics
+- `[blocksync]` Make minimum receive rate configurable via `min_recv_rate` in `[blocksync]` config section. Operators can tune this value based on their network's block size and block time characteristics
   ([\#5614](https://github.com/cometbft/cometbft/pull/5614))
 
 - `[mempool]` feat!(mempool): introduce app-mempool & follower-mode. Improve lib-p2p integration (project Krakatoa).